+++ packages/expo-ui/src/jetpack-compose/index.ts
@@ -49,6 +49,7 @@ export {
   type HorizontalPagerProps,
 } from './HorizontalPager';
 export * from './SearchBar';
+export * from './Snackbar';
 export * from './DockedSearchBar';
 export * from './HorizontalFloatingToolbar';
 export * from './FloatingActionButton';